Workaround to not update state in unmount components
- Book passage
Imagine the component mounts when the page is rendered for the first time in the browser.
In the componentDidMount() the component starts to make the request, but then the user navigates away from the page with this rendered component.
Then the App component unmounts, but there is still a pending request from componentDidMount() lifecycle method.
It will attempt to use this.setState() eventually in the then() or catch() block of the promise.
You will likely see the following warning on your command line, or in your browser's develop output:

Warning: Can only update a mounted or mounting component.

This usually means you called setState, replaceState or forceUpdate on an unmounted component. This is a no-op.